Read from secret files rather than ENV dictionary
