+++ .env.test
@@ -11,3 +11,6 @@ EXPORT_DATE_OVERRIDE=
 EPOCH_TIMESTAMP=1568059200
 
 ROLLBAR_TOKEN=
+
+AURORA_BOOTSTRAP_ROLLBAR_TOKEN_FILE=rollerbar_token
+AURORA_BOOTSTRAP_DB_CRED_FILE=db_pass
