+++ test/lib/rollerbar_token
@@ -1,0 +1,1 @@
+root
